# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ShortenerUrl API Restfull",
        "version": "1.0"
    },
    "paths": {
        "/api/shortenerUrl": {
            "post": {
                "tags": [
                    "shortenerUrl"
                ],
                "operationId": "6aa6ba51f0f023888b7107139e2cc52e",
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "original_url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "ShortenerUrl 1"
                                    },
                                    "original_url": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "http://example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store a newly created resource in storage.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shortenerUrl/{id}": {
            "delete": {
                "tags": [
                    "shortenerUrl"
                ],
                "operationId": "a0451e2f4218205eb088ac3948934193",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "The id of shortenerUrl",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store a newly created resource in storage.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "shortenerUrl"
                ],
                "operationId": "bfd98ab8451e06bde0b35b124fb275b7",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "The id of shortenerUrl",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "original_url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "ShortenerUrl 1"
                                    },
                                    "original_url": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "http://example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store a newly created resource in storage.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "shortenerUrl",
            "description": "shortenerUrl"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "name": "bearerAuth",
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}